refactor: Made auto adjusting grid logic into hook
